requirements.txt
app==0.0.1
bcrypt==3.1.7
DateTime==4.3
blis==0.2.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==7.1.2
cryptography==2.9.2
cymem==2.0.3
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.1
packaging==20.4
Pint==0.14
plac==0.9.6
preshed==2.0.1
psycopg2==2.8.5
pybase64==1.0.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2020.1
PyYAML==5.1.2
regex==2020.7.14
requests==2.24.0
six==1.15.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==1.0.2
thinc==7.0.8
tqdm==4.48.0
urllib3==1.25.10
wasabi==0.7.1
Werkzeug==1.0.1
zipp==3.1.0
